fix(renderer): clear the SonarCloud reliability findings in DDGI v2
The quality gate failed one condition, new_reliability_rating 3 (needs
1), on 7 bugs -- all in this PR's own code, all real:

cpp:S867 x6 -- `bool |= ImGui::Widget(...)` in the cascade settings.
Those six were the only `|=` uses in the file; the rest of the repo
writes `if (ImGui::X()) dirty = true;`, so matching the house style and
satisfying the rule are the same edit. Note the widget call stays
unconditional -- `dirty = ImGui::X() || dirty` would short-circuit and
skip drawing it.

cpp:S874 x1 -- `1 << glm::clamp(level, 0, 30)` shifts a signed operand
in CascadeSpacing. Now `1u << static_cast<u32>(...)`. Verified
value-identical across levels -5..40, both clamp saturations included,
so cascade spacing is unchanged.

+++ OloEngine/src/OloEngine/Renderer/DDGI/DDGICommon.h
@@ -394,7 +394,10 @@ namespace OloEngine::DDGI
     // cascade N covers 2^N times the extent at 2^N times the probe spacing.
     [[nodiscard("the cascade spacing is the only effect")]] inline glm::vec3 CascadeSpacing(const glm::vec3& baseSpacing, i32 level) noexcept
     {
-        return baseSpacing * static_cast<f32>(1 << glm::clamp(level, 0, 30));
+        // Unsigned shift: `1 << level` on a signed left operand is UB the moment
+        // the clamp is ever widened past 30, and there is no signedness to want
+        // here -- the result is a positive power of two on its way to f32.
+        return baseSpacing * static_cast<f32>(1u << static_cast<u32>(glm::clamp(level, 0, 30)));
     }
 
     // Camera-centred cascade. The window is placed so the camera sits as close
